requirements.txt
# Core dependencies
python-dotenv==1.0.0
datasets==2.16.1
huggingface-hub==0.20.3

# AI Agent APIs
openai==1.12.0
anthropic==0.18.1
google-generativeai==0.3.2

# Static analysis tools (free)
bandit==1.7.7
semgrep==1.60.0
pylint==3.0.3
flake8==7.0.0

# Code execution sandboxing
RestrictedPython==8.1

# Utilities
numpy==1.26.4
pandas==2.2.0
tqdm==4.66.1
requests==2.31.0
aiohttp==3.9.1

# Logging and monitoring
colorlog==6.8.2

# Web app and A2A protocol
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic-ai==0.0.14
a2a-sdk[http-server]==0.1.0
websockets==13.1
jinja2==3.1.4
